Fix: main을 Feat/modal-component로 병합하던 중 발생한 충돌 해결
